Implement Hand methods
This commit is where the magic happens, right now a lot of the tests are red and some are false-positives (because of no implementation).

As I code the Hand class, I keep my tests running and pay attention to those that are or that turn red. Slowly fixing everything until every tests runs green.